+++ tests/whitelist.py
@@ -9,5 +9,6 @@ reports.Table.add_col
 reports.Table.get_row
 reports.Table.set_row_order
 lab.tools.deprecated
+lab.tools.get_lab_path
 
 Call